+++ R/useful-items.R
@@ -885,7 +885,7 @@ bs4Stars <- function(maxstar = 5, grade, status = "warning") {
 #' @param title Jumbotron title.
 #' @param lead Jumbotron lead.
 #' @param href Jumbrotron external link.
-#' @param btn_name Jumbotron button name.
+#' @param btnName Jumbotron button name.
 #' @param status Jumbotron background color. "primary", "success", "warning", "danger" or "info".
 #'
 #' @author David Granjon, \email{dgranjon@@ymail.com}
@@ -920,7 +920,7 @@ bs4Stars <- function(maxstar = 5, grade, status = "warning") {
 #' }
 #'
 #' @export
-bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL, btn_name = "More",
+bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL, btnName = "More",
                          status = c("primary", "warning", "danger", "info", "success")) {
   
   status <- match.arg(status)
@@ -940,13 +940,13 @@ bs4Jumbotron <- function(..., title = NULL, lead = NULL, href = NULL, btn_name =
     shiny::tags$p(class = "lead", lead),
     shiny::tags$hr(class = "my-4"),
     shiny::tags$p(...),
-    if (!is.null(btn_name)) {
+    if (!is.null(btnName)) {
       shiny::tags$a(
         class = paste0("btn btn-", btnStatus, " btn-lg"),
         href = href,
         target = "_blank",
         role = "button",
-        btn_name
+        btnName
       )
     }
   )