+++ NEWS.md
@@ -12,6 +12,7 @@
 - `descriptionBlock()`: number_color, number_icon, right_border and margin_bottom become
 camel case parms (numberColor, numberIcon, rightBorder and marginBottom)
 - numberIcon in `descriptionBlock()` only need the name of the icon ('times') instead
+- `bs4Jumbotron()`: btn_name becomes btnName
 
 ## New features
 - New `bs4CardLayout()`: simplify the way to deal with `bs4Card()`!